+++ docs/release-notes.rst
@@ -4,7 +4,10 @@
 Release notes
 =============
 
-Release 0.9.9 (unreleased)
+Release 0.10.1 (unreleased)
+==========================
+
+Release 0.10.0
 ==========================
 - `PR 643 <https://github.com/uber/petastorm/pull/643>`_: Tests: fix the unit test of batched dataloader with in-memory cache.
 - `PR 662 <https://github.com/uber/petastorm/pull/662>`_: Reader: shuffle row groups before sharding (user may specify a seed).